NewsChefAPI/NewsChefAPI/requirements.txt
Django==1.10.2
djangorestframework==3.4.7
httpie==0.9.6
Pygments==2.1.3
requests==2.11.1
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
dj-database-url==0.4.0